+++ src/QueryBuilder/DDLQueryBuilderInterface.php
@@ -392,12 +392,14 @@ interface DDLQueryBuilderInterface
      * Builds an SQL statement for dropping a DB table.
      *
      * @param string $table The table to drop.
+     * @param bool $ifExists Do not throw an error if the table does not exist.
+     * @param bool $cascade Automatically drop objects that depend on the table.
      *
      * @return string The SQL statement for dropping a DB table.
      *
      * Note: The method will quote the `table` parameter before using it in the generated SQL.
      */
-    public function dropTable(string $table): string;
+    public function dropTable(string $table, bool $ifExists = false, bool $cascade = false): string;
 
     /**
      * Creates an SQL command for dropping a unique constraint.